requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
av==17.0.0
babel==2.18.0
blobfile==3.2.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
csvw==3.7.0
ctranslate2==4.7.1
datasets==4.7.0
decorator==5.2.1
dill==0.4.0
diskcache==5.6.3
dlinfo==2.0.0
dotenv==0.9.9
einops==0.8.2
einx==0.4.1
filelock==3.20.0
flatbuffers==25.12.19
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.4.2
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hyper-connections==0.4.9
idna==3.11
ipinfo==5.4.1
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kagglehub==1.0.0
kagglesdk==0.1.16
language-tags==1.2.0
lazy-loader==0.5
librosa==0.11.0
llama_cpp_python==0.3.16
llvmlite==0.46.0
local-attention==1.11.2
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
neucodec==0.0.5
neutts==1.2.1
numba==0.64.0
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.24.3
packaging==26.0
pandas==3.0.1
phonemizer==3.3.0
pillow==12.0.0
platformdirs==4.9.4
pooch==1.9.0
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pycryptodomex==3.23.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
resemble-perth==1.0.1
rfc3986==1.5.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
segments==2.4.0
sentencepiece==0.2.1
setuptools==70.2.0
six==1.17.0
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
srt==3.5.3
sympy==1.14.0
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0+cpu
torch-einops-utils==0.0.30
torchao==0.16.0
torchaudio==2.10.0+cpu
torchdata==0.11.0
torchtune==0.6.1
torchvision==0.25.0+cpu
tqdm==4.67.3
transformers==4.56.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
vector-quantize-pytorch==1.17.8
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
